Work on making Thruster an ABC, and creating a subclass
This is only "kinda" what I want it to be. It turns out that the property and classmethod decorators don't play well together... I have the properties being stored on a class-level, with instance-level properties that call on the class-level ones. This makes the ABC implementation kinda clunky, so it may be removed.